Fixed conversion from size_t to int warning.
Caused by convert_str which stored the result of the first call to MultiByteToWideChar in a size_t and then passed that to the second call which was expecting an int.
Also added an explicit length check to construct(const char* s, size_t len) before casting the original size_t to an int and changed expection specifications into comments.