Fixed bounds error in bstr_t::s_str().
The method was reading on character too many from the wrapped BSTR and writing the converted character to one byte after the std::string buffer.  While this mostly works because this will write the the NULL-terminator that mostly implementation put after the buffer it is not legal and could cause problems if the character happens not to be NULL (possible as BSTRs don't have to be NULL-terminated).